+++ packages/expo/src/async-require/hmr.ts
@@ -162,13 +162,15 @@ const HMRClient: HMRClientNativeInterface = {
     hmrClient = client;
 
     const fullBundleUrl = (() => {
-      if (document?.currentScript && 'src' in document.currentScript) {
-        return document.currentScript.src;
+      const currentScript = document?.currentScript;
+      const bundleUrl = new URL(
+        currentScript && 'src' in currentScript ? currentScript.src : location.href,
+        location.href
+      );
+
+      if (!bundleUrl.searchParams.has('platform')) {
+        bundleUrl.searchParams.set('platform', process.env.EXPO_OS ?? 'web');
       }
-
-      const bundleUrl = new URL(location.href);
-
-      bundleUrl.searchParams.set('platform', process.env.EXPO_OS ?? 'web');
 
       return bundleUrl.toString();
     })();